% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_weights_set.R
\name{connectivity_weights_set}
\alias{connectivity_weights_set}
\title{Persistent homology-based weight sets of functional connectivity}
\usage{
connectivity_weights_set(connectivity, na.rm = FALSE)
}
\arguments{
\item{connectivity}{a numeric N*N matrix or data frame, of which element represents weighted connectivity between nodes. N is the number of nodes. (e.g. Pearson correlation matrix with diagonal element of 1).}

\item{na.rm}{logical; whether remove NA values in connectivity matrix}
}
\value{
W : a set of all weights in connectivity matrix.

WB0: a set of added weights when the number of dimension 0 (connected components) topological features decreases by 1.

WB1: a set of added weights when the number of dimension 1 (cycles) topological features increases by 1.

WB10: a set of weights making cycles until all nodes are connected. It is a subset of WB1 and complement of WB11. i.e. WB10 U WB11 = WB1.

WB11: a set of weights making cycles after all nodes are connected. It is a subset of WB1 and complement of WB10. i.e. WB10 U WB11 = WB1.
}
\description{
\code{connectivity_weights_set} calculate dimension 0 (connected components) and dimension 1 (cycles) persistent homology and generate death weights of dimension 0 topological objects and birth weights of dimension 1 topological objects.
}
\examples{
a <- runif(45,-0.5,0.8)
A <- vectomat(a)
connectivity_weights_set(A)
}
