% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PH_meas.R
\name{PH_meas}
\alias{PH_meas}
\title{Calculate persistent homology-based functional connectivity measures}
\usage{
PH_meas(connectivity)
}
\arguments{
\item{connectivity}{a numeric N*N matrix or data frame, of which element represents weighted connectivity between nodes. N is the number of nodes. (e.g. Pearson correlation matrix with diagonal element of 1).}
}
\value{
BS: Backbone Strength, BD: Backbone Dispersion, CS: Cycle strength
}
\description{
\code{PH_meas} calculate persistent homology-based functional connectivity measures
}
\examples{
a <- runif(19900,-0.5,0.8)
A <- vectomat(a)
PH_meas(A)

}
\references{
Persistent Homology-based Functional Connectivity Explains Cognitive Ability: Life-span Study,
Hyunnam Ryu, Christian G. Habeck, Yaakov Stern, and Seonjoo Lee,
bioRxiv 2022.10.17.512619; doi: https://doi.org/10.1101/2022.10.17.512619
}
