% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TStoConn.R
\name{TStoConn}
\alias{TStoConn}
\title{Convert Time Series Matrix to Connectivity Matrix for PH calculation}
\usage{
TStoConn(ts, is.abs = FALSE)
}
\arguments{
\item{ts}{#time-points(T)*#ROIs(N) timeseries matrix}

\item{is.abs}{logical. TRUE returns a connectivity matrix with absolute correlation coefficient values. Default is FALSE.}
}
\value{
K*K symmetric matrix (K is the number of survived ROIs (or voxels) after screening)
}
\description{
\code{TStoConn} converts #time-points(T)*#ROIs(N) timeseries matrix or data frame to N*N symmetric connectivity matrix
for PH calculation while while removing ROIs(or voxels) whose values are all zero or all the same.
}
\examples{
TS <- matrix(runif(100*5,-1,1),nrow=100,ncol=5)
TStoConn(TS, is.abs=TRUE)
}
