% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uppermat.R
\name{uppermat}
\alias{uppermat}
\title{Extract upper diagonal elements from a symmetric matrix}
\usage{
uppermat(mat)
}
\arguments{
\item{mat}{symmetric matrix}
}
\value{
a numeric vector
}
\description{
\code{uppermat} extract a vector with N*(N-1)/2 upper diagonal values from a N*N symmetric matrix.
}
\examples{
a <- runif(6,-0.5,0.8)
A <- vectomat(a)
uppermat(A)

}
