% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectomat.R
\name{vectomat}
\alias{vectomat}
\title{Convert Vector to symmetric Matrix}
\usage{
vectomat(vec, val.diag = 1)
}
\arguments{
\item{vec}{a vector with N*(N-1)/2 numeric values}

\item{val.diag}{a value which is assigned to the diagonal. Default is 1.}
}
\value{
N*N symmetric matrix
}
\description{
\code{vectomat} converts vector with N*(N-1)/2 values to N*N symmetric matrix.
}
\examples{
a <- runif(6,-0.5,0.8)
vectomat(a)
}
