% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PH_meas.R
\name{PH_meas}
\alias{PH_meas}
\title{Calculate persistent homology-based functional connectivity measures}
\usage{
PH_meas(connectivity)
}
\arguments{
\item{connectivity}{a numeric N*N matrix or data frame, of which element represents weighted connectivity between nodes. N is the number of nodes. (e.g. Pearson correlation matrix with diagonal element of 1).}
}
\value{
BS: Backbone Strength, BD: Backbone Dispersion, CS: Cycle strength, CF: Cycle frequency, aBD: adjusted Backbone Dispersion
}
\description{
\code{PH_meas} calculate persistent homology-based functional connectivity measures
}
\examples{
a <- runif(19900,-0.5,0.8)
A <- vectomat(a)
PH_meas(A)

}
\references{
Ryu, H., Habeck, C., Stern, Y., & Lee, S. (2023). Persistent homology-based functional connectivity and its association with cognitive ability: Life-span study. Human Brain Mapping, 44( 9), 3669– 3683. https://doi.org/10.1002/hbm.26304
}
